Fix group views functionality after consolidation
