--- src/components/groupColorsView/groupColorsView.view.yaml
@@ -1,90 +1,0 @@
-elementName: rvn-group-colors-view
-
-viewDataSchema:
-  type: object
-  properties:
-    searchQuery:
-      type: string
-
-propsSchema:
-  type: object
-  properties:
-    flatGroups:
-      type: array
-      description: Array of color groups to display
-    selectedItemId:
-      type: string
-      description: ID of currently selected item
-
-refs:
-  search-input:
-    eventListeners:
-      input-change:
-        handler: handleSearchInput
-  group-*:
-    eventListeners:
-      click:
-        handler: handleGroupClick
-  color-item-*:
-    eventListeners:
-      click:
-        handler: handleColorItemClick
-      dblclick:
-        handler: handleColorItemDoubleClick
-  add-color-button-*:
-    eventListeners:
-      click:
-        handler: handleAddColorClick
-  add-color-dialog:
-    eventListeners:
-      close:
-        handler: handleCloseDialog
-      submit-form:
-        handler: handleSubmitForm
-  color-form:
-    eventListeners:
-      action-click:
-        handler: handleFormActionClick
-
-template:
-  - 'rtgl-view w=f h=100vh g=lg style="overflow-y: scroll;"':
-      - rtgl-view g=lg wh=f:
-          - 'rtgl-view h=48 w=f bgc=bg bwb=xs ph=md av=c style="position: sticky; top: 0px; z-index: 1000;"':
-              - rtgl-input#search-input s=sm placeholder="Search colors..." value=${searchQuery}:
-          - $if flatGroups.length > 0:
-              - $for group in flatGroups:
-                  - rtgl-view w=f ph=md:
-                      - 'rtgl-view#group-${group.id} d=h w=f av=c bgc=bg cur=p style="position: sticky; top: 48px;"':
-                          - $if group.isCollapsed:
-                              - rtgl-svg wh=16 svg=folderArrowRight:
-                            $else:
-                              - rtgl-svg wh=16 svg=folderArrowDown:
-                          - rtgl-view d=h av=c g=md:
-                              - rtgl-svg wh=16 svg=folder:
-                              - rtgl-text: ${group.fullLabel}
-                      - $if !group.isCollapsed:
-                          - rtgl-view w=f mb=md p=sm:
-                              - rtgl-view d=h fw=w g=md:
-                                  - $if group.hasChildren:
-                                      - $for item in group.children:
-                                          - 'rtgl-view#color-item-${item.id} br=md g=sm bw=xs cur=p style="overflow: hidden; ${item.selectedStyle}"':
-                                              - 'rtgl-view w=200 h=150 style="background-color: ${item.hex}"':
-                                              - rtgl-text s=sm c=mu-fg m=md ellipsis w="calc(200px - 2 * var(--spacing-md))": ${item.name}
-                                      - "rtgl-view#add-color-button-${group.id} w=200 h=150 av=c ah=c g=md bw=xs bc=bo br=md cur=p":
-                                          - rtgl-svg svg=plus wh=24:
-                                          - rtgl-text: Add Color
-                                    $else:
-                                      - "rtgl-view#add-color-button-${group.id} w=200 h=150 av=c ah=c g=md bw=xs bc=bo br=md cur=p":
-                                          - rtgl-svg svg=plus wh=24:
-                                          - rtgl-text: Add Color
-            $elif searchQuery:
-              - rtgl-view w=f h=400 av=c ah=c:
-                  - rtgl-text s=lg c=mu-fg: No colors found matching "${searchQuery}"
-            $else:
-              - rtgl-view w=f h=f av=c ah=c:
-                  - rtgl-text s=lg c=mu-fg: No data
-
-              - rtgl-view h=64:
-
-      - $if flatGroups.length > 0:
-          - rtgl-view h=33vh:
