--- src/components/groupComponentsView/groupComponentsView.store.js
@@ -1,106 +1,0 @@
-export const INITIAL_STATE = Object.freeze({
-  collapsedIds: [],
-  searchQuery: "",
-  isDialogOpen: false,
-  currentGroupId: null,
-});
-
-export const toggleGroupCollapse = (state, groupId) => {
-  const index = state.collapsedIds.indexOf(groupId);
-  if (index > -1) {
-    state.collapsedIds.splice(index, 1);
-  } else {
-    state.collapsedIds.push(groupId);
-  }
-};
-
-export const setSearchQuery = (state, query) => {
-  state.searchQuery = query;
-};
-
-export const showDialog = (state, groupId) => {
-  state.isDialogOpen = true;
-  state.currentGroupId = groupId;
-};
-
-export const hideDialog = (state) => {
-  state.isDialogOpen = false;
-  state.currentGroupId = null;
-};
-
-export const selectCurrentGroupId = ({ state }) => state.currentGroupId;
-
-export const toViewData = ({ state, props }) => {
-  const selectedItemId = props.selectedItemId;
-  const searchQuery = state.searchQuery.toLowerCase();
-
-  // Helper function to check if an item matches the search query
-  const matchesSearch = (item) => {
-    if (!searchQuery) return true;
-
-    const name = (item.name || "").toLowerCase();
-    const description = (item.description || "").toLowerCase();
-
-    return name.includes(searchQuery) || description.includes(searchQuery);
-  };
-
-  // Apply collapsed state and search filtering to flatGroups
-  const flatGroups = (props.flatGroups || [])
-    .map((group) => {
-      // Filter children based on search query
-      const filteredChildren = (group.children || []).filter(matchesSearch);
-
-      // Only show groups that have matching children or if there's no search query
-      const hasMatchingChildren = filteredChildren.length > 0;
-      const shouldShowGroup = !searchQuery || hasMatchingChildren;
-
-      return {
-        ...group,
-        isCollapsed: state.collapsedIds.includes(group.id),
-        children: state.collapsedIds.includes(group.id)
-          ? []
-          : filteredChildren.map((item) => ({
-              ...item,
-              selectedStyle:
-                item.id === selectedItemId
-                  ? "outline: 2px solid var(--color-pr); outline-offset: 2px;"
-                  : "",
-            })),
-        hasChildren: filteredChildren.length > 0,
-        shouldDisplay: shouldShowGroup,
-      };
-    })
-    .filter((group) => group.shouldDisplay);
-
-  return {
-    flatGroups,
-    selectedItemId: props.selectedItemId,
-    searchQuery: state.searchQuery,
-    isDialogOpen: state.isDialogOpen,
-    defaultValues: {
-      name: "",
-    },
-    form: {
-      title: "Add Component",
-      fields: [
-        {
-          name: "name",
-          inputType: "inputText",
-          label: "Component Name",
-          description: "Enter the component name",
-          required: true,
-        },
-      ],
-      actions: {
-        layout: "",
-        buttons: [
-          {
-            id: "submit",
-            variant: "pr",
-            content: "Add Component",
-          },
-        ],
-      },
-    },
-  };
-};
