--- src/components/groupComponentsView/groupComponentsView.view.yaml
@@ -1,89 +1,0 @@
-elementName: rvn-group-components-view
-
-viewDataSchema:
-  type: object
-  properties:
-    searchQuery:
-      type: string
-
-propsSchema:
-  type: object
-  properties:
-    flatGroups:
-      type: array
-      description: Array of component groups to display
-    selectedItemId:
-      type: string
-      description: ID of currently selected item
-
-refs:
-  search-input:
-    eventListeners:
-      input-change:
-        handler: handleSearchInput
-  add-component-button-*:
-    eventListeners:
-      click:
-        handler: handleAddComponentClick
-  add-component-dialog:
-    eventListeners:
-      close:
-        handler: handleCloseDialog
-  component-form:
-    eventListeners:
-      action-click:
-        handler: handleFormActionClick
-  group-*:
-    eventListeners:
-      click:
-        handler: handleGroupClick
-  component-item-*:
-    eventListeners:
-      click:
-        handler: handleComponentItemClick
-      dblclick:
-        handler: handleComponentItemDoubleClick
-
-template:
-  - 'rtgl-view w=f h=100vh g=lg style="overflow-y: scroll;"':
-    - rtgl-view g=lg w=f h=f:
-      - 'rtgl-view h=48 w=f bgc=bg bwb=xs ph=md av=c style="position: sticky; top: 0px; z-index: 1000;"':
-        - rtgl-input#search-input s=sm placeholder="Search components..." value=${searchQuery}:
-        # TODO implement zoom controls (increase/decrease component size)
-      - $if flatGroups.length > 0:
-          - $for group in flatGroups:
-            - rtgl-view w=f ph=md:
-              - 'rtgl-view#group-${group.id} d=h w=f av=c bgc=bg cur=p style="position: sticky; top: 48px;"':
-                - $if group.isCollapsed:
-                    - rtgl-svg wh=16 svg=folderArrowRight:
-                  $else:
-                    - rtgl-svg wh=16 svg=folderArrowDown:
-                - rtgl-view d=h av=c g=md:
-                  - rtgl-svg wh=16 svg=folder:
-                  - rtgl-text: ${group.fullLabel}
-              - $if !group.isCollapsed:
-                  - rtgl-view w=f mb=md p=sm:
-                    - rtgl-view d=h fw=w w=f g=md:
-                      - $if group.hasChildren:
-                          - $for item in group.children:
-                              - 'rtgl-view#component-item-${item.id} w=200 h=150 ah=c av=c bw=xs br=sm cur=p style="${item.selectedStyle}"':
-                                - rtgl-view av=c ah=c:
-                                  - rtgl-text s=md: ${item.name}
-                                  - rtgl-text s=sm c=mu-fg: Component
-                      - 'rtgl-view#add-component-button-${group.id} w=200 h=150 ah=c av=c bw=xs br=sm cur=p':
-                        - rtgl-text: Add Component
-        $elif searchQuery:
-          - rtgl-view w=f h=400 av=c ah=c:
-            - rtgl-text s=lg c=mu-fg: No components found matching "${searchQuery}"
-        $else:
-          - rtgl-view w=f h=f av=c ah=c:
-              - rtgl-text s=lg c=mu-fg: No data
-
-              - rtgl-view h=64:
-
-      - $if flatGroups.length > 0:
-          - rtgl-view h=33vh:
-
-  - rtgl-dialog#add-component-dialog ?open=${isDialogOpen}:
-    - rtgl-view slot=content g=md:
-      - rtgl-form#component-form .defaultValues=defaultValues .form=form w=f:
