--- src/components/groupLayoutsView/groupLayoutsView.view.yaml
@@ -1,90 +1,0 @@
-elementName: rvn-group-layouts-view
-
-viewDataSchema:
-  type: object
-  properties:
-    searchQuery:
-      type: string
-
-propsSchema:
-  type: object
-  properties:
-    flatGroups:
-      type: array
-      description: Array of layout groups to display
-    selectedItemId:
-      type: string
-      description: ID of currently selected item
-
-refs:
-  search-input:
-    eventListeners:
-      input-change:
-        handler: handleSearchInput
-  add-layout-button-*:
-    eventListeners:
-      click:
-        handler: handleAddLayoutClick
-  add-layout-dialog:
-    eventListeners:
-      close:
-        handler: handleCloseDialog
-  layout-form:
-    eventListeners:
-      action-click:
-        handler: handleFormActionClick
-  group-*:
-    eventListeners:
-      click:
-        handler: handleGroupClick
-  layout-item-*:
-    eventListeners:
-      click:
-        handler: handleLayoutItemClick
-      dblclick:
-        handler: handleLayoutItemDoubleClick
-
-template:
-  - 'rtgl-view w=f h=100vh g=lg style="overflow-y: scroll;"':
-    - rtgl-view g=lg w=f h=f:
-      - 'rtgl-view h=48 w=f bgc=bg bwb=xs ph=md av=c style="position: sticky; top: 0px; z-index: 1000;"':
-        - rtgl-input#search-input s=sm placeholder="Search layouts..." value=${searchQuery}:
-        # TODO implment zoom controls (increase/decrease layout size)
-      - $if flatGroups.length > 0:
-          - $for group in flatGroups:
-            - rtgl-view w=f ph=md:
-              - 'rtgl-view#group-${group.id} d=h w=f av=c bgc=bg cur=p style="position: sticky; top: 48px;"':
-                - $if group.isCollapsed:
-                    - rtgl-svg wh=16 svg=folderArrowRight:
-                  $else:
-                    - rtgl-svg wh=16 svg=folderArrowDown:
-                - rtgl-view d=h av=c g=md:
-                  - rtgl-svg wh=16 svg=folder:
-                  - rtgl-text: ${group.fullLabel}
-              - $if !group.isCollapsed:
-                  - rtgl-view w=f mb=md p=sm:
-                    - rtgl-view d=h fw=w w=f g=md:
-                      - $if group.hasChildren:
-                          - $for item in group.children:
-                              - 'rtgl-view#layout-item-${item.id} w=200 h=150 ah=c av=c bw=xs br=sm cur=p style="${item.selectedStyle}"':
-                                - rtgl-view av=c ah=c:
-                                  - rtgl-text s=md: ${item.name}
-                                  - rtgl-text s=sm c=mu-fg: ${item.layoutTypeDisplay}
-                      - 'rtgl-view#add-layout-button-${group.id} w=200 h=150 av=c ah=c g=md bw=xs bc=bo br=md cur=p':
-                        - rtgl-svg svg=plus wh=24:
-                        - rtgl-text: Add Layout
-        $elif searchQuery:
-          - rtgl-view w=f h=400 av=c ah=c:
-            - rtgl-text s=lg c=mu-fg: No layouts found matching "${searchQuery}"
-        $else:
-          - rtgl-view w=f h=f av=c ah=c:
-              - rtgl-text s=lg c=mu-fg: No data
-
-              - rtgl-view h=64:
-
-      - $if flatGroups.length > 0:
-          - rtgl-view h=33vh:
-
-  - rtgl-dialog#add-layout-dialog ?open=${isDialogOpen} s=md:
-    - rtgl-view slot=content g=md:
-      - rtgl-form#layout-form .defaultValues=defaultValues .form=form w=f:
