+++ src/pages/characters/characters.view.yaml
@@ -52,25 +52,25 @@ refs:
 
 template:
   - rtgl-view d=h w=f h=f:
-    - rtgl-view d=h h=f w=f:
-      - rvn-resizable-panel#file-explorer w=300 min-w=200 max-w=500 resize-side="right":
-        - rtgl-view slot="content" w=f h=f:
-          - rvn-resource-types .resourceCategory=resourceCategory .selectedResourceId=selectedResourceId:
-          - rvn-file-explorer .items=flatItems .repositoryTarget=repositoryTarget:
-      - rtgl-view flex=1 h=f:
-        - rvn-group-resources-view#groupview .resourceType=resourceType .flatGroups=flatGroups .selectedItemId=selectedItemId .searchQuery=searchQuery .searchPlaceholder=searchPlaceholder:
-      - rvn-resizable-panel panel-type=detail-panel w=270 min-w=200 max-w=500 resize-side="left":
-        - rtgl-view wh=f slot="content":
-          - $if selectedItemId:
-            - rtgl-form#detail-form key=${selectedItemId} w=f h=f .form=form .context=context .defaultValues=defaultValues:
-            $else:
-              - rtgl-view wh=f av=c ah=c bgc=mu-fg:
-                  - rtgl-text s=md c=mu-fg: No selection
-    - rtgl-dialog#add-character-dialog ?open=${isDialogOpen} s=md:
-      - rtgl-view slot=content g=md:
-        - rtgl-form#character-form key=${isDialogOpen} .defaultValues=dialogDefaultValues .form=dialogForm w=f:
-          - $if avatarFileId:
-              - rvn-file-image#avatar-image fileId=${avatarFileId} wh=120 slot=avatar-slot cur=p br=md:
-            $else:
-              - rtgl-view#avatar-placeholder slot=avatar-slot wh=120 av=c ah=c bgc=bg bc=bo bw=xs br=md cur=p:
-                - rtgl-text c=mu-fg s=sm: Click to Upload
+      - rtgl-view d=h h=f w=f:
+          - rvn-resizable-panel#file-explorer w=300 min-w=200 max-w=500 resize-side="right":
+              - rtgl-view slot="content" w=f h=f:
+                  - rvn-resource-types .resourceCategory=resourceCategory .selectedResourceId=selectedResourceId:
+                  - rvn-file-explorer .items=flatItems .repositoryTarget=repositoryTarget:
+          - rtgl-view flex=1 h=f:
+              - rvn-group-resources-view#groupview full-width-item .resourceType=resourceType .flatGroups=flatGroups .selectedItemId=selectedItemId .searchQuery=searchQuery .searchPlaceholder=searchPlaceholder:
+          - rvn-resizable-panel panel-type=detail-panel w=270 min-w=200 max-w=500 resize-side="left":
+              - rtgl-view wh=f slot="content":
+                  - $if selectedItemId:
+                      - rtgl-form#detail-form key=${selectedItemId} w=f h=f .form=form .context=context .defaultValues=defaultValues:
+                    $else:
+                      - rtgl-view wh=f av=c ah=c bgc=mu-fg:
+                          - rtgl-text s=md c=mu-fg: No selection
+      - rtgl-dialog#add-character-dialog ?open=${isDialogOpen} s=md:
+          - rtgl-view slot=content g=md:
+              - rtgl-form#character-form key=${isDialogOpen} .defaultValues=dialogDefaultValues .form=dialogForm w=f:
+                  - $if avatarFileId:
+                      - rvn-file-image#avatar-image fileId=${avatarFileId} wh=120 slot=avatar-slot cur=p br=md:
+                    $else:
+                      - rtgl-view#avatar-placeholder slot=avatar-slot wh=120 av=c ah=c bgc=bg bc=bo bw=xs br=md cur=p:
+                          - rtgl-text c=mu-fg s=sm: Click to Upload
