+++ src/pages/colors/colors.store.js
@@ -26,6 +26,8 @@ export const INITIAL_STATE = Object.freeze({
   editItemId: null,
   isAddDialogOpen: false,
   targetGroupId: null,
+  searchQuery: "",
+  collapsedIds: [],
   context: {
     colorImage: {
       src: "",
@@ -74,6 +76,19 @@ export const closeAddDialog = (state) => {
   state.targetGroupId = null;
 };
 
+export const setSearchQuery = (state, query) => {
+  state.searchQuery = query;
+};
+
+export const toggleGroupCollapse = (state, groupId) => {
+  const index = state.collapsedIds.indexOf(groupId);
+  if (index > -1) {
+    state.collapsedIds.splice(index, 1);
+  } else {
+    state.collapsedIds.push(groupId);
+  }
+};
+
 export const selectSelectedItem = ({ state }) => {
   if (!state.selectedItemId) return null;
   const flatItems = toFlatItems(state.colorsData);
@@ -84,7 +99,7 @@ export const selectSelectedItemId = ({ state }) => state.selectedItemId;
 
 export const toViewData = ({ state, props }, payload) => {
   const flatItems = toFlatItems(state.colorsData);
-  const flatGroups = toFlatGroups(state.colorsData);
+  const rawFlatGroups = toFlatGroups(state.colorsData);
 
   // Get selected item details
   const selectedItem = state.selectedItemId
@@ -99,6 +114,49 @@ export const toViewData = ({ state, props }, payload) => {
       rgb: hexToRgb(selectedItem.hex) || "",
     };
   }
+
+  // Apply search filter
+  const searchQuery = state.searchQuery.toLowerCase().trim();
+  let filteredGroups = rawFlatGroups;
+
+  if (searchQuery) {
+    filteredGroups = rawFlatGroups
+      .map((group) => {
+        const filteredChildren = (group.children || []).filter((item) => {
+          const name = (item.name || "").toLowerCase();
+          const hex = (item.hex || "").toLowerCase();
+          return name.includes(searchQuery) || hex.includes(searchQuery);
+        });
+
+        const groupName = (group.name || "").toLowerCase();
+        const shouldIncludeGroup =
+          filteredChildren.length > 0 || groupName.includes(searchQuery);
+
+        return shouldIncludeGroup
+          ? {
+              ...group,
+              children: filteredChildren,
+              hasChildren: filteredChildren.length > 0,
+            }
+          : null;
+      })
+      .filter(Boolean);
+  }
+
+  // Apply collapsed state and selection styling
+  const flatGroups = filteredGroups.map((group) => ({
+    ...group,
+    isCollapsed: state.collapsedIds.includes(group.id),
+    children: state.collapsedIds.includes(group.id)
+      ? []
+      : (group.children || []).map((item) => ({
+          ...item,
+          selectedStyle:
+            item.id === state.selectedItemId
+              ? "outline: 2px solid var(--color-pr); outline-offset: 2px;"
+              : "",
+        })),
+  }));
 
   // Get edit item details
   const editItem = state.editItemId
@@ -199,5 +257,8 @@ export const toViewData = ({ state, props }, payload) => {
     isAddDialogOpen: state.isAddDialogOpen,
     addDefaultValues,
     addForm,
+    searchQuery: state.searchQuery,
+    resourceType: "colors",
+    searchPlaceholder: "Search colors...",
   };
 };
