--- src/pages/components/components.view.yaml
@@ -1,40 +1,0 @@
-elementName: rvn-components
-
-viewDataSchema:
-  type: object
-  properties: {}
-
-refs:
-  file-explorer:
-    eventListeners:
-      data-changed:
-        handler: handleDataChanged
-  groupview:
-    eventListeners:
-      image-item-click:
-        handler: handleImageItemClick
-      files-uploaded:
-        handler: handleDragDropFileSelected
-      component-created:
-        handler: handleComponentCreated
-  detail-form:
-    eventListeners:
-      form-change:
-        handler: handleFormChange
-
-template: 
-  - rtgl-view d=h w=f h=f: 
-    - rtgl-view d=h h=f w=f:
-      - rvn-resizable-panel#file-explorer w=300 min-w=200 max-w=500 resize-side="right":
-        - rtgl-view slot="content" w=f h=f:
-          - rvn-resource-types .resourceCategory=resourceCategory .selectedResourceId=selectedResourceId:
-          - rvn-file-explorer .items=flatItems .repositoryTarget=repositoryTarget:
-      - rtgl-view flex=1 h=f:
-        - rvn-group-components-view#groupview .flatGroups=flatGroups .selectedItemId=selectedItemId:
-      - rvn-resizable-panel panel-type=detail-panel w=270 min-w=200 max-w=500 resize-side="left":
-        - rtgl-view wh=f slot="content":
-          - $if selectedItemId:
-            - rtgl-form#detail-form key=${selectedItemId} w=f h=f .form=form .defaultValues=defaultValues .fieldResources=fieldResources:
-            $else:
-              - rtgl-view wh=f av=c ah=c bgc=mu-fg:
-                  - rtgl-text s=md c=mu-fg: No selection
