+++ src/pages/layouts/layouts.view.yaml
@@ -2,7 +2,9 @@ elementName: rvn-layouts
 
 viewDataSchema:
   type: object
-  properties: {}
+  properties:
+    searchQuery:
+      type: string
 
 refs:
   file-explorer:
@@ -11,30 +13,46 @@ refs:
         handler: handleDataChanged
   groupview:
     eventListeners:
-      image-item-click:
+      item-click:
         handler: handleImageItemClick
+      item-dblclick:
+        handler: handleItemDoubleClick
       files-uploaded:
         handler: handleDragDropFileSelected
-      layout-created:
-        handler: handleLayoutCreated
+      add-layout-click:
+        handler: handleAddLayoutClick
+      search-input:
+        handler: handleSearchInput
+      group-toggle:
+        handler: handleGroupToggle
   detail-form:
     eventListeners:
       form-change:
         handler: handleFormChange
+  add-layout-dialog:
+    eventListeners:
+      close:
+        handler: handleAddDialogClose
+  layout-form:
+    eventListeners:
+      action-click:
+        handler: handleLayoutFormActionClick
 
-template: 
-  - rtgl-view d=h w=f h=f: 
-    - rtgl-view d=h h=f w=f:
-      - rvn-resizable-panel#file-explorer w=300 min-w=200 max-w=500 resize-side="right":
-        - rtgl-view slot="content" w=f h=f:
-          - rvn-resource-types .resourceCategory=resourceCategory .selectedResourceId=selectedResourceId:
-          - rvn-file-explorer .items=flatItems .repositoryTarget=repositoryTarget .contextMenuItems=contextMenuItems .emptyContextMenuItems=emptyContextMenuItems:
-      - rtgl-view flex=1 h=f:
-        - rvn-group-layouts-view#groupview .flatGroups=flatGroups .selectedItemId=selectedItemId:
-      - rvn-resizable-panel panel-type=detail-panel w=270 min-w=200 max-w=500 resize-side="left":
-        - rtgl-view wh=f slot="content":
-          - $if selectedItemId:
-            - rtgl-form#detail-form key=${selectedItemId} w=f h=f .form=form .defaultValues=defaultValues .fieldResources=fieldResources:
-            $else:
-              - rtgl-view wh=f av=c ah=c bgc=mu-fg:
-                  - rtgl-text s=md c=mu-fg: No selection
+template:
+  - rtgl-view d=h w=f h=f:
+      - rtgl-view d=h h=f w=f:
+          - rvn-resizable-panel#file-explorer w=300 min-w=200 max-w=500 resize-side="right":
+              - rtgl-view slot="content" w=f h=f:
+                  - rvn-resource-types .resourceCategory=resourceCategory .selectedResourceId=selectedResourceId:
+                  - rvn-file-explorer .items=flatItems .repositoryTarget=repositoryTarget .contextMenuItems=contextMenuItems .emptyContextMenuItems=emptyContextMenuItems:
+          - rtgl-view flex=1 h=f:
+              - rvn-group-resources-view#groupview .resourceType=resourceType .flatGroups=flatGroups .selectedItemId=selectedItemId .searchQuery=searchQuery .searchPlaceholder=searchPlaceholder:
+          - rvn-resizable-panel panel-type=detail-panel w=270 min-w=200 max-w=500 resize-side="left":
+              - rtgl-view wh=f slot="content":
+                  - $if selectedItemId:
+                      - rtgl-form#detail-form key=${selectedItemId} w=f h=f .form=form .defaultValues=defaultValues .fieldResources=fieldResources:
+                    $else:
+                      - rtgl-view wh=f av=c ah=c bgc=mu-fg:
+                          - rtgl-text s=md c=mu-fg: No selection
+  - rtgl-dialog#add-layout-dialog ?open=${isAddDialogOpen} s=md:
+      - rtgl-form#layout-form slot=content .defaultValues=layoutFormDefaults .form=layoutForm w=f:
