+++ src/pages/transforms/transforms.store.js
@@ -17,6 +17,8 @@ const form = {
 export const INITIAL_STATE = Object.freeze({
   transformData: { tree: [], items: {} },
   selectedItemId: null,
+  searchQuery: "",
+  collapsedIds: [],
   contextMenuItems: [
     { label: "New Folder", type: "item", value: "new-item" },
     { label: "Duplicate", type: "item", value: "duplicate-item" },
@@ -36,13 +38,26 @@ export const setSelectedItemId = (state, itemId) => {
   state.selectedItemId = itemId;
 };
 
+export const setSearchQuery = (state, query) => {
+  state.searchQuery = query;
+};
+
+export const toggleGroupCollapse = (state, groupId) => {
+  const index = state.collapsedIds.indexOf(groupId);
+  if (index > -1) {
+    state.collapsedIds.splice(index, 1);
+  } else {
+    state.collapsedIds.push(groupId);
+  }
+};
+
 export const selectSelectedItemId = ({ state }) => {
   return state.selectedItemId;
 };
 
 export const toViewData = ({ state, props }, payload) => {
   const flatItems = toFlatItems(state.transformData);
-  const flatGroups = toFlatGroups(state.transformData);
+  const rawFlatGroups = toFlatGroups(state.transformData);
 
   // Get selected item details
   const selectedItem = state.selectedItemId
@@ -63,6 +78,48 @@ export const toViewData = ({ state, props }, payload) => {
     };
   }
 
+  // Apply search filter
+  const searchQuery = state.searchQuery.toLowerCase().trim();
+  let filteredGroups = rawFlatGroups;
+
+  if (searchQuery) {
+    filteredGroups = rawFlatGroups
+      .map((group) => {
+        const filteredChildren = (group.children || []).filter((item) => {
+          const name = (item.name || "").toLowerCase();
+          return name.includes(searchQuery);
+        });
+
+        const groupName = (group.name || "").toLowerCase();
+        const shouldIncludeGroup =
+          filteredChildren.length > 0 || groupName.includes(searchQuery);
+
+        return shouldIncludeGroup
+          ? {
+              ...group,
+              children: filteredChildren,
+              hasChildren: filteredChildren.length > 0,
+            }
+          : null;
+      })
+      .filter(Boolean);
+  }
+
+  // Apply collapsed state and selection styling
+  const flatGroups = filteredGroups.map((group) => ({
+    ...group,
+    isCollapsed: state.collapsedIds.includes(group.id),
+    children: state.collapsedIds.includes(group.id)
+      ? []
+      : (group.children || []).map((item) => ({
+          ...item,
+          selectedStyle:
+            item.id === state.selectedItemId
+              ? "outline: 2px solid var(--color-pr); outline-offset: 2px;"
+              : "",
+        })),
+  }));
+
   return {
     flatItems,
     flatGroups,
@@ -74,5 +131,8 @@ export const toViewData = ({ state, props }, payload) => {
     emptyContextMenuItems: state.emptyContextMenuItems,
     form,
     defaultValues,
+    searchQuery: state.searchQuery,
+    resourceType: "transforms",
+    searchPlaceholder: "Search transforms...",
   };
 };
