+++ src/pages/typography/typography.view.yaml
@@ -2,7 +2,9 @@ elementName: rvn-typography
 
 viewDataSchema:
   type: object
-  properties: {}
+  properties:
+    searchQuery:
+      type: string
 
 refs:
   file-explorer:
@@ -11,8 +13,10 @@ refs:
         handler: handleDataChanged
   groupview:
     eventListeners:
-      typography-item-click:
+      item-click:
         handler: handleTypographyItemClick
+      item-dblclick:
+        handler: handleTypographyItemDoubleClick
       files-uploaded:
         handler: handleDragDropFileSelected
       typography-created:
@@ -21,8 +25,10 @@ refs:
         handler: handleTypographyUpdated
       add-typography-click:
         handler: handleAddTypographyClick
-      typography-item-double-click:
-        handler: handleTypographyItemDoubleClick
+      search-input:
+        handler: handleSearchInput
+      group-toggle:
+        handler: handleGroupToggle
   detail-form:
     eventListeners:
       form-change:
@@ -68,7 +74,7 @@ template:
           - rvn-resource-types .resourceCategory=resourceCategory .selectedResourceId=selectedResourceId:
           - rvn-file-explorer .items=flatItems .repositoryTarget=repositoryTarget .contextMenuItems=contextMenuItems .emptyContextMenuItems=emptyContextMenuItems:
       - rtgl-view flex=1 h=f:
-        - rvn-group-typography-view#groupview .flatGroups=flatGroups .selectedItemId=selectedItemId .colorsData=colorsData .fontsData=fontsData:
+        - rvn-group-resources-view#groupview .resourceType=resourceType .flatGroups=flatGroups .selectedItemId=selectedItemId .searchQuery=searchQuery .searchPlaceholder=searchPlaceholder:
       - rvn-resizable-panel panel-type=detail-panel w=270 min-w=200 max-w=500 resize-side="left":
         - rtgl-view wh=f slot="content":
           - $if selectedItemId:
